server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.3
httpx>=0.27.0
aiosmtplib==3.0.1
email-validator==2.1.0
websockets==12.0
Pillow>=10.2.0
aiofiles==23.2.1